fixup! refactor: Move base.php to a proper class
